Avoid to cast to long in get_system_cpu_times()'s Py_BuildValue()

+++ psutil/_psutil_bsd.c
@@ -439,8 +439,8 @@ static PyObject* get_system_cpu_times(PyObject* self, PyObject* args)
         return NULL;
     }
 
-    return Py_BuildValue("(lll)", (long)cpu_time[CP_USER],
-                                  (long)cpu_time[CP_SYS],
-                                  (long)cpu_time[CP_IDLE]
+    return Py_BuildValue("(lll)", cpu_time[CP_USER],
+                                  cpu_time[CP_SYS],
+                                  cpu_time[CP_IDLE]
                         );
 }
